+++ ext/EnsembleKalmanFiltersExt.jl
@@ -5,7 +5,22 @@ using Ensembles:
 using EnsembleKalmanFilters: EnsembleKalmanFilters, EnKF
 
 function Ensembles.assimilate_data(
-    filter::EnKF, ensemble, ensemble_obs_clean, ensemble_obs_noisy, y_obs, log_data
+    filter::EnKF, ensemble::AbstractEnsemble, ensemble_obs::AbstractEnsemble, y_obs, log_data
+)
+    X_matrix = EnsembleKalmanFilters.assimilate_data(
+        filter,
+        Float64.(get_ensemble_matrix(ensemble)),
+        Float64.(get_ensemble_matrix(ensemble_obs)),
+        get_member_vector(ensemble_obs_clean, y_obs),
+        log_data,
+    )
+    members = get_ensemble_dicts(ensemble, X_matrix)
+    posterior = Ensemble(members, ensemble.state_keys)
+    return posterior
+end
+
+function Ensembles.assimilate_data(
+    filter::EnKF, ensemble::AbstractEnsemble, ensemble_obs_clean::AbstractEnsemble, ensemble_obs_noisy::AbstractEnsemble, y_obs, log_data
 )
     X_matrix = EnsembleKalmanFilters.assimilate_data(
         filter,